Take notes on the post-boot RAM layout where XBIOS-4 /should be/ in RAM.
It looks like its copied to 0x36c0, and there should be an 0xCD included
in front of it.

At 0x0F34 there's an 0xFF, which is RST 38. Now, either I'm copying it
wrong, or something's explicitly putting a CD there somehow beforehand
during copying or initialisation.

In any case that explains why my disassembly is so messy, I need to
actually somehow call out what's going on here and feed it into
z80dasm so it doesn't generate a trashfire disassembly.